+++ src/components/sdk/MeetingControls/AudioOutputControl.tsx
@@ -1,16 +1,16 @@
 // Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import React from 'react';
+import { isOptionActive, supportsSetSinkId } from '../../../utils/device-utils';
 
 import { ControlBarButton } from '../../ui/ControlBar/ControlBarItem';
-import { Sound } from '../../ui/icons';
-import { useMeetingManager } from '../../../providers/MeetingProvider';
-import { useAudioOutputs } from '../../../providers/DevicesProvider';
-import { useLocalAudioOutput } from '../../../providers/LocalAudioOutputProvider';
-import { isOptionActive } from '../../../utils/device-utils';
 import { DeviceType } from '../../../types';
 import { PopOverItemProps } from '../../ui/PopOver/PopOverItem';
+import React from 'react';
+import { Sound } from '../../ui/icons';
+import { useAudioOutputs } from '../../../providers/DevicesProvider';
+import { useLocalAudioOutput } from '../../../providers/LocalAudioOutputProvider';
+import { useMeetingManager } from '../../../providers/MeetingProvider';
 
 interface Props {
   /** The label that will be shown for audio output speaker control, it defaults to `Speaker`. */
@@ -21,13 +21,17 @@ const AudioOutputControl: React.FC<Props> = ({ label = 'Speaker' }) => {
   const meetingManager = useMeetingManager();
   const { devices, selectedDevice } = useAudioOutputs();
   const { isAudioOn, toggleAudio } = useLocalAudioOutput();
+  const audioOutputOnClick = async (deviceId: string): Promise<void> => {
+    if (supportsSetSinkId()) {
+      await meetingManager.selectAudioOutputDevice(deviceId);
+    }
+  }
 
   const dropdownOptions: PopOverItemProps[] = devices.map(
     (device: DeviceType) => ({
       children: <span>{device.label}</span>,
       checked: isOptionActive(selectedDevice, device.deviceId),
-      onClick: (): Promise<void> =>
-        meetingManager.selectAudioOutputDevice(device.deviceId)
+      onClick: (): Promise<void> => audioOutputOnClick(device.deviceId),
     })
   );