+++ src/components/ui/icons/Caution/Caution.stories.tsx
@@ -1,12 +1,12 @@
 // Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import React from 'react';
+import { select, text } from '@storybook/addon-knobs';
 
-import { text } from '@storybook/addon-knobs';
 import Caution from './';
 import CautionIconDocs from './Caution.mdx';
 import Flex from '../../Flex';
+import React from 'react';
 
 export default {
   title: 'UI Components/Icons/Caution',
@@ -18,4 +18,8 @@ export default {
   component: Caution
 };
 
-export const _Caution = () => <Flex layout="fill-space-centered"><Caution width={text('width', '2rem')} /></Flex>;
+export const _Caution = () => (
+  <Flex layout="fill-space-centered">
+    <Caution width={text('width', '2rem')} variant={select('displayStyle', ['default', 'fill-warning', 'fill-error'], 'default')} />
+  </Flex>
+);