+++ src/hooks/sdk/useSelectAudioOutputDevice.tsx
@@ -1,8 +1,8 @@
 // Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import { supportsSetSinkId } from '../../utils/device-utils';
 import { useCallback } from 'react';
-
 import { useMeetingManager } from '../../providers/MeetingProvider';
 
 export const useSelectAudioOutputDevice = () => {